Clarify that some model classes are only used internally by AWApiReachHelper
- Make AWReach, AWReachDetail, AWRapid nested classes within AWApiReachHelper
- Move files under a new group in API
